connect to s3

+++ app/views.py
@@ -74,14 +74,11 @@ def colony_recognition():
     return render_template("public/colony_recognition.html")
 
 
-
 """
 \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
     RUN PIE Algo IN FLASK
 \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-
 """
-
 
 def allowed_image(filename):
 
@@ -102,21 +99,56 @@ def allowed_image_filesize(filesize):
     else:
         return False
 
+
 @app.route("/upload-image", methods=["GET", "POST"])
 
 def upload_file_s3(file_name, bucket, object_name=None):
     if object_name is None:
         object_name = file_name
 
-    #upload the file
-    s3_client = boto3.client('s3')
+    #upload the file to s3
+    config = Config(signature_version=botocore.UNSIGNED)
+    s3_client = boto3.client('s3', config=config)
+    s3_resource = boto3.resource('s3')
+    
+    success = False
+    
+    # check if bucket is created OK
     try:
-        response = s3_client.upload_file(file_name, bucket, object_name)
+        bucket = s3_resource.Bucket(bucket)
     except ClientError as e:
-        logging.error(e)
-        return False
-    return True
+        bucket = None
 
+    # Check if client is created OK, and if not etag(file ID) set to be empty
+    try:
+        head = s3_client.head_object(Bucket= bucket, Key=object_name)
+    except ClientError:
+        etag=''
+    else:
+        etag = head['ETag'].strip('"')
+
+    # create bucket object (resource representing s3 object)
+    try:
+        s3_obj = bucket.Object(filename)
+    except ClientError, AttributeError:
+        s3_obj = None
+
+    # upload the file to the bucket
+    try:
+        s3_obj.upload_fileobj(file_name)
+    except ClientError, AttributeError:
+        pass # then success is still false
+    else:
+        try:
+            # return the object only if its entity tag(etag) is different from the one specificed, otw return 304
+            s3_obj.wait_until_exists(IfNoneMatch=etag) 
+        except WaiterError as e:
+            logging.error(e)
+        else:
+            head = s3_client.head_object(Bucket=bucket, Key=file_name)
+            success = head['ContentLength']
+
+    return success
 
 def upload_image():
 
@@ -145,13 +177,12 @@ def upload_image():
                     save_path = os.path.join(app.config["IMAGE_UPLOADS"], filename)
                     file.save(save_path)
 
-                    # upload input file to S3
-                    upload_file_s3(save_path, "pie-data", filename)
+                    # upload input file to S3 & granting public access to s3
+                    upload_file_s3(save_path, "pie-data", filename, ExtraArgs={'ACL': 'public-read'})
 
-
-                    # run analysis
+                    # run analysis and store the processed file to S3
                     process_file(input_src=save_path)
-                    
+
                     # remove input file from server
                     os.remove(save_path)
                     print("file removed from the server")
@@ -167,50 +198,49 @@ def upload_image():
     return render_template("public/upload_image.html")
 
 
-def move_file(source, destination):
-    files = os.listdir(source)
-    for file in files:
-        shutil.move(os.path.join(source, file), os.path.join(destination, file))
+# def move_file(source, destination):
+#     files = os.listdir(source)
+#     for file in files:
+#         shutil.move(os.path.join(source, file), os.path.join(destination, file))
 
 
 def process_file(input_src):
+
     # locate files
-    
-    output_dst = app.config["CLIENT_IMAGES"] #/Users/hyunjung/Projects/FlaskProject/app/static/client/img
-    #input_dst = '/Users/hyunjung/MATLAB-Drive/simple_code_test_ims'
-    #move_file(input_src, input_dst)
+    output_dst = app.config["CLIENT_IMAGES"] # /Users/hyunjung/Projects/FlaskProject/app/static/client/img
+    # input_dst = '/Users/hyunjung/MATLAB-Drive/simple_code_test_ims'
+    # move_file(input_src, input_dst)
 
     colony_mask, colony_property_df = read_and_run_analysis(
         hole_fill_area=np.inf, cleanup=False, max_proportion_exposed_edge=0.75,
         save_extra_info=False, image_type="brightfield", input_im_path=input_src, output_path=output_dst)
-    
+
+    # upload processed files to s3 bucket
     for output_file in os.listdir(output_dst):
-        output_file = os
+        output_file = os.path.join(app.config["CLIENT_IMAGES"], output_file)
 
-        upload_file_s3(output_file, "pie-data")
-        os.remove(output_file)
-    
-    
-
+        # upload file to s3 
+        success = upload_file_s3(output_file, "pie-data")
+        if success != False:
+            os.remove(output_file)
+            print("file successfully uploaded to s3")
+        else:
+            abort(404)
 
     # move output file
-    #output_src = '/Users/hyunjung/MATLAB-Drive/simple_code_test_ims_output/boundary_ims'
-    #output_dst = app.config["CLIENT_IMAGES"]
-    #move_file(output_src, output_dst)
+    # output_src = '/Users/hyunjung/MATLAB-Drive/simple_code_test_ims_output/boundary_ims'
+    # output_dst = app.config["CLIENT_IMAGES"]
+    # move_file(output_src, output_dst)
 
 @app.route("/get-image/<filename>")
-
 
 def get_image(filename):
 
     try:
-        print(app.config["CLIENT_IMAGES"])
-        print(filename)
-
-
+        url = "https://s3.ap-northeast-2.amazonaws.com/"
 
         return send_from_directory(
-            app.config["CLIENT_IMAGES"],
+            url,
             filename=filename,
             as_attachment=True
         )
